requirements.txt
tensorflow
pillow
numpy
opencv-python
